fix overflow handling in hashagg CE
hashagg currently assumes each group by is independent, so it multiplies all keys's distinct to estimate number of rows in the final output. This can cause overflow say tpch 1g q10.

Also, enforce all CE returns ulong instead long to reflect the facts cardinality shall not be negative.

+++ qpmodel/Catalog.cs
@@ -187,7 +187,7 @@ namespace qpmodel
             {
                 Catalog.systable_.CreateTable($"T{i}", cols, null);
                 var stat = new ColumnStat();
-                stat.n_rows_ = 1 + i * 10;
+                stat.n_rows_ = (ulong)(1 + i * 10);
                 Catalog.sysstat_.AddOrUpdate($"T{i}", "i", stat);
             }
         }
